refactor: common으로 이동함에 따라 테스트 코드 변경
