+++ commerce/src/test/java/store/_0982/commerce/application/grouppurchase/ParticipateServiceConcurrencyTest.java
@@ -11,10 +11,13 @@ import org.springframework.test.context.DynamicPropertySource;
 import org.testcontainers.containers.GenericContainer;
 import org.testcontainers.junit.jupiter.Container;
 import store._0982.commerce.domain.grouppurchase.GroupPurchaseRepository;
-import store._0982.commerce.domain.grouppurchase.GroupPurchaseStatus;
 import store._0982.commerce.domain.product.ProductRepository;
 import store._0982.commerce.exception.CustomErrorCode;
 import store._0982.commerce.support.BaseConcurrencyTest;
+import store._0982.common.domain.grouppurchase.GroupPurchase;
+import store._0982.common.domain.grouppurchase.GroupPurchaseStatus;
+import store._0982.common.domain.product.Product;
+import store._0982.common.domain.product.ProductCategory;
 import store._0982.common.exception.CustomException;
 
 import java.time.OffsetDateTime;
