+++ commerce/src/test/java/store/_0982/commerce/application/product/ProductServiceTest.java
@@ -16,6 +16,8 @@ import store._0982.commerce.application.product.dto.*;
 import store._0982.commerce.domain.grouppurchase.GroupPurchaseRepository;
 import store._0982.commerce.domain.product.ProductRepository;
 import store._0982.commerce.domain.product.ProductVectorRepository;
+import store._0982.common.domain.product.Product;
+import store._0982.common.domain.product.ProductCategory;
 import store._0982.common.dto.PageResponse;
 import store._0982.common.exception.CustomException;
